fix: after lesson
