+++ AutoTestsSelenium/PageObjects/AbstractStudentAuthorizedPage.cs
@@ -2,7 +2,7 @@
 {
     public abstract class AbstractStudentAuthorizedPage : AbstractAuthorizedPage
     {
-        public IWebElement ButtonHomeworksSideBar => _driver.FindElement(By.XPath($"//*[text()='Домашние задания']/.."));
+        public IWebElement ButtonHomeworksSideBar => GetButtonHomeworksSideBar();
         public IWebElement ButtonLessonsSideBar => GetButtonLessonsSideBar();
 
         protected AbstractStudentAuthorizedPage()
@@ -23,6 +23,12 @@
         {
             WebDriverWait webDriverWait = new WebDriverWait(_driver, TimeSpan.FromSeconds(0.5));
             return webDriverWait.Until(ExpectedConditions.ElementExists(By.XPath($"//*[text()='Занятия']/..")));
+        }
+        
+        private IWebElement GetButtonHomeworksSideBar()
+        {
+            WebDriverWait webDriverWait = new WebDriverWait(_driver, TimeSpan.FromSeconds(0.5));
+            return webDriverWait.Until(ExpectedConditions.ElementExists(By.XPath($"//*[text()='Домашние задания']/..")));
         }
     }
 }
