+++ AutoTestsSelenium/PageObjects/AuthorizationUnauthorizedPage.cs
@@ -33,8 +33,6 @@
         public void ClickEnterButton()
         {
             ButtonEnter.Click();
-            WebDriverWait webDriverWait = new WebDriverWait(_driver, TimeSpan.FromSeconds(1));
-            webDriverWait.Until(ExpectedConditions.ElementExists(By.XPath($"//*[text()='Уведомления']/..")));
         }
 
         public void ClickCancelButton()
